+++ CHANGES.md
@@ -44,6 +44,7 @@ Changes in 12.5
 - JS remove inline code from program in Export.php & Export.js
 - JS remove inline code from program in Courses.php & Courses.js
 - JS remove inline code from program in MassDrops.php & MassDrops.js
+- JS remove inline code from program in MassRequests.php & MassRequests.js
 
 Changes in 12.4.3
 -----------------